requirements.txt
aiofiles==24.1.0
asyncmy==0.2.9
greenlet==3.1.1
ispyb==10.2.3
mysql-connector-python==9.0.0
pika==1.3.2
pillow==10.4.0
pip==23.0.1
setuptools==65.5.0
SQLAlchemy==1.4.54
tabulate==0.9.0
tqdm==4.66.5
